requirements.txt
pyspark==3.5.1
delta-spark==3.2.0
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1